checking methods' parameters

+++ src/main/java/jm/task/core/jdbc/service/UserServiceImpl.java
@@ -19,10 +19,16 @@ public class UserServiceImpl implements UserService {
     }
 
     public void saveUser(String name, String lastName, byte age) {
+        if (name != null || name.length() == 0 || lastName == null || lastName.length() == 0 || age < 0) {
+            return;
+        }
         userDao.saveUser(name, lastName, age);
     }
 
     public void removeUserById(long id) {
+        if (id < 0) {
+            return;
+        }
         userDao.removeUserById(id);
     }
